rootfs/usr/src/requirements.txt
pyyaml==6.0.3
pyserial==3.5
paho-mqtt==2.1.0
pydantic==2.10.6
